Now it works but...
It works, but it isn’t properly storing the key into the table, but
vertex is being stored.

ALSO, it isn’t skipping over whitespace and new lines when reading in
past the first line. So the edges aren’t getting properly stores. Seg
faults on first attempt to store an edge (because it’s trying to store
two null characters, which it can’t hash)